+++ controller/scenarios/processPlaceOrderByStaff.ts
@@ -135,7 +135,7 @@ export async function main(organizationIdentifier: string, durationInMillisecond
         tel: customer.telephone,
         gender: '0',
         age: '',
-        address: ''
+        address: 'JP'
     };
     customerContact = await placeOrderTransactions.setCustomerContact({
         transactionId: transaction.id,