+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Changed
 
+- install ttts-domain@13.3.0
+
 ### Deprecated
 
 ### Removed